refactor generate random cluster value function
